Allow strings to be keys in the supplied data
If your template is "{{foo}}" then `%{"foo" => "data"}` should work just as
well as `%{foo: "data"}`. It's a bit of an edge case, but if you want to
interpolate user-supplied data into the template, turning the keys into atoms
is inadvisable as they never get GC'ed. If your keys are all strings, the
atoms won't even be tried.